# spec/swagger.yaml
swagger: '2.0'
info:
  contact:
    email: nitish.malhotra@gmail.com
  description: |
    This is a RESTful API for the vegeta load-testing utility. Vegeta is a versatile HTTP load testing tool built out of a need to drill HTTP services with a constant request rate.
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  title: Vegeta REST API
  version: 1.0.0
host: virtserver.swaggerhub.com
paths:
  '/attack':
    post:
      summary: Attack a target server
      tags:
        - attack
      operationId: postAttack
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: Attack specification object. See https://github.com/tsenart/vegeta#attack-command
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Attack'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AttackResponse'
        # '400':
        #   description: Bad Request
        #   schema:
        #     $ref: '#/definitions/Error'
        # '500':
        #   description: Internal Server Error
        #   schema:
        #     $ref: '#/definitions/Error'
    get:
      summary: Get a list of all attack details
      tags:
        - attack
      operationId: getAttacks
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AttackResponseList'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Attack:
    properties:
      target:
        $ref: '#/definitions/Target'
      body:
        description: Request body encoded as a json string
        type: string
      connections:
        default: 10000
        description: Max open idle connections per target host
        format: int32
        type: integer
      duration:
        description: Duration of the test
        example: 10s
        format: duration 
        type: string
      h2c:
        default: true
        description: Send HTTP/2 requests without TLS encryption
        type: boolean
      headers:
        description: Request headers as a string array.
        example:
          - 'Accept: text/html'
          - 'Last-Modified: Sat, 29 Dec 2018 07:28:00 GMT'
        items:
          type: string
        type: array
      http2:
        default: true
        description: Send HTTP/2 requests when supported by the server
        type: boolean
      insecure:
        description: Specifies whether to ignore invalid server TLS certificates.
        type: boolean
      keepalive:
        description: Specifies whether to reuse TCP connections between HTTP requests.
        type: boolean
      laddr:
        description: Specifies the local IP address to be used.
        format: ipv4
        type: string
      max-body:
        description: Specifies the maximum number of bytes to capture from the body of each response. Remaining unread bytes will be fully read but discarded. Set to -1 for no limit.
        example: 10M
        type: string
      rate:
        default: 50
        description: 'Specifies the request rate per time unit to issue against the targets. The actual request rate can vary slightly due to things like garbage collection, but overall it should stay very close to the specified. If no time unit is provided, 1s is used.'
        example: 50
        format: int
        type: integer 
      redirects:
        description: 'Specifies the max number of redirects followed on each request. The default is 10. When the value is -1, redirects are not followed but the response is marked as successful.'
        format: int
        type: integer
      resolvers:
        description: Specifies custom DNS resolver addresses to use for name resolution instead of the ones configured by the operating system. Works only on non Windows systems.
        type: string
      timeout:
        description: Specifies the timeout for each request. The default is 0 which disables timeouts.
        example: 10s
        format: time
        type: string
      workers:
        default: 10
        description: Specifies the initial number of workers used in the attack. The actual number of workers will increase if necessary in order to sustain the requested rate.
        format: int
        type: integer
      cert:
        description: TLS client PEM encoded certificate base64 encoded
        example: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUYyekNDQk1PZ0F3SUJBZ0lRTWo4SGpnd2VYa2J3TUxWSk9OMGJnVEFOQmdrcWhraUc5dzBCQVFzRkFEQ0IKcERFTE1Ba0dBMVVFQmhNQ1ZWTXhIVEFiQmdOVkJBb1RGRk41YldGdWRHVmpJRU52Y25CdmNtRjBhVzl1TVI4dwpIUVlEVlFRTEV4WkdUMUlnVkVWVFZDQlFWVkpRVDFORlV5QlBUa3haTVI4d0hRWURWUVFMRXhaVGVXMWhiblJsCll5QlVjblZ6ZENCT1pYUjNiM0pyTVRRd01nWURWUVFERXl0VGVXMWhiblJsWXlCRGJHRnpjeUF6SUZObFkzVnkKWlNCVFpYSjJaWElnVkVWVFZDQkRRU0F0SUVjME1CNFhEVEUyTURZeU9EQXdNREF3TUZvWERURTRNRFl5T1RJegpOVGsxT1Zvd2daSXhDekFKQmdOVkJBWVRBbFZUTVJNd0VRWURWUVFJREFwRFlXeHBabTl5Ym1saE1SWXdGQVlEClZRUUhEQTFOYjNWdWRHbGhiaUJXYVdWM01SMHdHd1lEVlFRS0RCUlRlVzFoYm5SbFl5QkRiM0p3YjNKaGRHbHYKYmpFZ01CNEdBMVVFQ3d3WFYxTlRJQzBnVkdWamFHNXBZMkZzSUZOMWNIQnZjblF4RlRBVEJnTlZCQU1NRENvdQpZbUowWlhOMExtNWxkRENDQVNJd0RRWUpLb1pJaHZjTkFRRUJCUUFEZ2dFUEFEQ0NBUW9DZ2dFQkFOcUJLa082CmUvUDcvUGFXRlgyZHQrZHJBS2hBbi9MUkRwNjJFRjQxYU1VL1hZbmxhMThiNUZ6VWMyZGhFbWUxN1Zudjh0WDUKQnJiby8zaHIrZHFQaEpkZnI2TmJVdnZsc0JHVXJscnBlZzFBUkhpdGEyY1BvYjdCRFAyalFvV0R0TTM2cndaQgp2a2d6Nys4QlB2WnFraXdxeVpFZTBoMGw3ZG1Ob3pNdHQ1ODdwZkxieTg2K3RmUjFyTFpIbnh3K0RMZS8rZ3BwClhpSHRUTXZDNm12R21sb3VZbVg5OHBiL2kyUG55WG1vaWloU3FrbndNNzRvTTN6RERHM0x1MHc4eENCQTlaLy8KTjByY1JHU2llYnJuZTAxS2dmb0ZSRVhBTVBMRXlyQzhTMG9zcHVNcTB5Yk1pbjVGcjA3UDZubkMrMUtibHpFTQpMbFJvR1A4cDQ4R2l3OUVDQXdFQUFFWEFNUExFZ2dJVE1Dd0dBMVVkRVFRbE1DT0NFM2QzZHk1emRXSXhMbUppCmRHVnpkQzV1WlhTQ0RDb3VZbUowWkVYQU1QTEVkREFKQmdOVkhSTUVBakFBTUE0R0ExVWREd0VCL3dRRUF3SUYKb0RBZEJnTlZIU1VFRmpBVUJnZ3JCZ0VGQlFjREFRWUlLd1lCQlFVSEF3SXdZUVlEVlIwZ0JGb3dXREJXQmdabgpnUXdCQWdJd1REQWpCZ2dyQmdFRkJRY0NBUllYYUhSMGNITTZMeTlrTG5ONWJXTmlMbU52YlM5amNITXdKUVlJCkt3WUJCUVVIQWdJd0dRd1hhSFIwY0hNNkx5OWtMbk41YldOaUxtTnZiUzl5Y0dFd0h3WURWUjBqQkJnd0ZvQVUKTkk5VXRUOEtIMUs2bkxKbDdicUxDR2NaNEFRd0t3WURWUjBmQkNRd0lqQWdvQjZnSElZYWFIUjBjRG92TDNOegpMbk41YldOaUxtTnZiUzl6Y3k1amNtd3dWd1lJS3dZQkJRVUhBUUVFU3pCSk1COEdDQ3NHQVFVRkJ6QUJoaE5vCmRIUndPaTh2YzNNdWMzbHRZMlF1WTI5dE1DWUdDQ3NHQVFVRkJ6QUNoaHBvZEhSd09pOHZjM011YzNsdFkySXUKWTI5dEwzTnpMbU55ZERBU0JnTXJaVTBFQ3pBSkFnRUNBZ0VBQWdFQU1JR0tCZ29yQmdFRUFkWjVBZ1FDQkh3RQplZ0I0QUhZQXozR0FDb1EwQWpRWW5velNXamNEVXZ1WisyZllEOHR3QjJjNFlnQnFyRmtBQUFGVmwxak1SZ0FBCkJBTUFSekJGQWlCY3RINHJnL3ZkNVVDdG5ZT2FGYVhJNEhKNlM2Qms4VDlHUHBIRU5EUDBOQUloQU5lTy8yK2gKLzFhSnFQT2U1dUdmRHFpcFM1d1dyVVVPYUpwVTlmdUhvREpRTUEwR0NTcUdTSWIzRFFFQkN3VUFBNElCQVFBdQpRL01mWXJEUVA0UHFuekVVb0FtNndhVloyaG0vM0g4MHNRWERyTjFPNTBaTnZqSE5zanp3ZENFdTFid0ZmYTZiCjBiOFA0eTRuS2E0YU9sMC9tSmxxL0F3ZnJ4MHVDODFVSU1Md21YMm1aNWRiYVhIS0VaaDJ0SENobkR3aFpHUEUKS1ZaSjBLeE9BMTJDVFJNRU9HNVhIVGU3WWwrd09RV204aDBreTBEWStGTVQ1QWdjQ282SU14TWJ5eXhvdFF0RQovOERtd3RaUXl0QTJ5cXRaV3EwNzY1dDZQQ0pTYm5LNnpwMGFMVFN3WVpWaWoxQkNETVlsZWlaY0R2SU42SnYvCkVsbkR3cnhzMUNzTXdoN3pZOHdCOGdjN0dIcTYzQkJXMWhDd3NEeDJndURDRW1iSmEra3R2N0VCejJCZ2lMNlYKWmcrUXFJRnowWVNEUUpmRk1UaQotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0tCg==
        type: string
      key:
        description: TLS cleint PEM encoded private key base64 encoded
        example: LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRRE5qRnNMN0R5d1A0bm8KaElpZUo4dHBXZHJET3NSVmhKbHV1cEJzdDJIL1RYUVBML2dybVlPMlQwNGlZV0NEb3RzSk95MTU2Zld1WldieApIaElXMmJ3T0d2T2ljRVpBY1c1MDdIMVlTY0hzbHFyMlp2WVB3cENUQmdaMStXTHJjaTRFQ1NpcXFuZXh2V2UvCkhHeXh4YTdnNzNCWHNqeUJxdTF1ZllGdVRlYVdBZmNGRHBpTC84YVVNbVNYZ1FtTUJVc05VclhQVklyOW5kV1QKaU9iNzVUWWM3eHF1YjFpMUFTZFJ5WVczbFlOZzZxZElaL1pWYnVaVXRwZml1L2hMaTloODZLMzZtUExFTFk2WgpvNFVyMUsxNUkyZ3BDYmcxV3dwSDdzenVwUDNpTU9vYUhrQUl5TVhvK1Z2V2RUamZJY0cyaTczWmlSQWY4Z2lECnY1aytsK1g3QWdNQkFBRUNnZ0VBY2cwTEI5YVE1Z3FiYnRUN0tTWlRpNlU3NlUvV1dNMDJoV1luZE90Y0NOamwKS1Bnbzc2aENtUW53Qk11YlRlOHpNakFQejdHdGpKTGZSN0t6dnRJWnNvbnAzWndHbndOUmtyY1Y1Y01uUXl0bQp1VW53QmxhaStiRkNMSHV3RTR4TFZCbnc0cUdyd0llajVUclVaTitYWlYvMktmTFA5L0xydzM4SUxqZ3pWYy9KCnZQeVUrWUhsWWJkMTA0am4xbGlHUkRCWk10cGF0alJ1YmNYeWlzTVFaR294UzVJSWh4N0dYV0ZXd1dJa1RSZ2IKN3VlcnFaR1BRY3oxOXNyVkhySDBnT0MxdElhMXBmaUdBZGs3SDlqZFFGQVJqUVhFcTVzRlcrY2tsdVRPZTZVZQpuTHZlcXdzaUJ6Y2E5ZUdHWnZ4a3N0aERnTUk0SmZzcjl2cTA0WVUrb1FLQmdRRDZtYzFSVUhtVVR2VUJhNmc5CnZNM1V3eHpnNk5zV1lkdytFVForN01QckdxdVhRWXBSeU1yaFFDM0pLUXV3M0hNaHNydjI5RVVXMFgwOTFrWncKd20yOTBlSjlzamovci93YXZPZytWbXEvbWh5STZKMnJxTUF6ZnNYZkRFSlNidUhlRXZGL0FhMG8wZmtMZG5JWAp5SjRsSlBjczBUcERyeS9QcmdGN2R0UUhwUUtCZ1FEUitoQ1RoOWhSWThKTjhpZ1I5ZWZTYzUxa3NIN1NyNlNjCmErZDBJbmJualdrdDFlZ3VKQjBzUmZ6TldWYWR1UzNTRXNuTTl3VlBUazkxY3lib2VwREd0RGxPOTlYSVd4V0UKcjE4aG1WSXV6Sm5oZnlxdHEySzF1OTF0OS9JcGRrU3lURkZQUGZOSUIrMkVJQ3UwbGIrSnBwU3Bvd2lJb1dVeQpYd2ZXNXBYRkh3S0JnQ2Z2a3JTK3FtRmtJODlTSk1HUjhRdFMvS3NxOEo4a2dVQ3k1WWhjbXV4QWw4L0c5aEFLCmU2cGgyK29PekhvbTU0SWJSYkVaUkZsYWtTYmxtdHdsWCsrOEh1Z0JHdXhFVTF5L2RoTlpHakhta1YyUUZIQmkKdUI1OHVYUE9ORk9Hc0IvL3pOUjhvK2U1Ym9MUklXV0E1eExzVFNlclFDbkJKVkZuWFgwYmRyYXhBb0dCQUo4UApTL0ZHRFpLUi9FV3BjN3N2U1QzOXdMUGVreU9zVkFqK0lHVzBHZGlNSWFuTUNVYXlxcnE0QmFuVXpDMU1jcTVZCjl5OWh6RTdvSm43MURQczhwV0V3Y3Qwa0ZuamdycXRGWWM0RTQrRjVzZlBFcmlMQzJsS0tqbEZZYXNRSHlFQTgKV2JlcVdSa3VrRGp2V2N4QlI4ZlJQWHBIZXhadEZFaVUvazVIZkJySEFvR0JBUEo3YlFOV1JEQzlLWVNSaGw4WQpvUzYwT0FFNVZrV29IVzI3N2RDUVU1MEpGL1YreStxRlZkWTBWVjd4MVVlbXVjekRFUFFIUXE5MXMybzRDdCs1CmpSQWcwS2xVRlFjQ0tEK2FCd28rZmZiLzJ6OVlJb1NXT3dsS0VrMWxMcFZ1RGYxU041NlkyQUw5ZGF2bEhCUFAKcTQrZXNwMGp1QkpnekxaeFM1R2l1S3BWCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0K
        type: string
      root-certs:
        description: 'Specifies the trusted TLS root CAs certificate strings as a comma separated list. If unspecified, the default system CAs certificates will be used.'
        type: string
    required:
      - duration
      - rate
      - target
    type: object
  AttackResponse:
    properties:
      id:
        description: The attack ID used to fetch status and reports.
        type: string
      status:
        description: The attack status
        enum:
          - scheduled
          - running
          - canceled
          - completed
          - failed
        type: string
    type: object
  AttackResponseList:
    items:
      $ref: '#/definitions/AttackResponse'
    type: array
  Target:
    properties:
      url:
        type: string
        example: virtserver.swaggerhub.com/api/v1/attack
      method:
        default: GET
        description: Method
        enum:
          - GET
          - POST
          - PUT
          - HEAD
          - DELETE
        type: string
      scheme:
        default: http
        enum:
          - http
          - https
        type: string
    type: object
...
  Error:
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - code
      - message
...
schemes:
 - https
 - http
basePath: /api/v1